Move AdminUnitBuilder to opengever.testing.
